+++ src/components/Instructions.tsx
@@ -2,7 +2,7 @@ import download from '../assets/download.webp'
 import  Icons  from './Icons'
 const Instructions = (): JSX.Element => {
     return (
-      <section id="instructions" className='lg:py-12 py-8 px-4 bg-gradient-to-t from-transparent sm:px-6 lg:px-8 dark:bg-zinc-950'>
+      <section id="instructions" className='lg:py-12 py-8 px-4 bg-gradient-to-t from-transparent sm:px-6 lg:px-8 bg-zinc-950'>
         <div className="container">
             <div className="relative grid gap-[17px] sm:grid-cols-2 xl:grid-cols-[repeat(15,minmax(0,1fr))]">
                 <div className="relative xl:col-span-5">